removed types.py + moved utils class functions  into utils submodule
