--- pythereum/types.py
@@ -1,61 +1,0 @@
-import collections.abc
-import numbers
-from typing import (
-    Any,
-)
-from .dclasses import HexStr
-
-bytes_types = (bytes, bytearray)
-integer_types = (int,)
-text_types = (str,)
-string_types = (bytes, str, bytearray)
-
-
-def is_integer(value: Any) -> bool:
-    return isinstance(value, integer_types) and not isinstance(value, bool)
-
-
-def is_bytes(value: Any) -> bool:
-    return isinstance(value, bytes_types)
-
-
-def is_text(value: Any) -> bool:
-    return isinstance(value, text_types)
-
-
-def is_string(value: Any) -> bool:
-    return isinstance(value, string_types)
-
-
-def is_boolean(value: Any) -> bool:
-    return isinstance(value, bool)
-
-
-def is_dict(obj: Any) -> bool:
-    return isinstance(obj, collections.abc.Mapping)
-
-
-def is_list_like(obj: Any) -> bool:
-    return not is_string(obj) and isinstance(obj, collections.abc.Sequence)
-
-
-def is_list(obj: Any) -> bool:
-    return isinstance(obj, list)
-
-
-def is_tuple(obj: Any) -> bool:
-    return isinstance(obj, tuple)
-
-
-def is_null(obj: Any) -> bool:
-    return obj is None
-
-
-def is_number(obj: Any) -> bool:
-    return isinstance(obj, numbers.Number)
-
-def is_hexstr(obj: Any) -> bool:
-    return isinstance(obj, HexStr)
-
-def is_float(obj: Any) -> bool:
-    return isinstance(obj, float)
